// src/app/@core/models/form9s/sign/sign-form.model.ts
export class SignFormModel {
    id?: number = 0;
    formID?: string = null;
    stepID?: number = 0;
    stepName?: string = null;
    signOptionID?: number = 0;
    signOption?: string = null;
    signResultID?: number = 0;
    signContent?: string = null;
    status: number = 0;
    arrivalDate?: Date = null;
    fileGroupID?: string = null;
    stepDescription?: string = null;
    formData?: any = null;
}